requirements.txt
azure-ai-documentintelligence==1.0.0b4
azure-core>=1.29.1
azure-identity
azure-keyvault-secrets
flask
dotenv
pdf2image
